+++ src/igvjs/ui/genericContainer.js
@@ -23,10 +23,10 @@
  * THE SOFTWARE.
  */
 
-import $ from "../vendor/jquery-1.12.4.js";
+import $ from '../vendor/jquery-1.12.4.js';
+import { DomUtils } from '../../../node_modules/igv-ui/dist/igv-ui.js';
 import { makeDraggable } from "./draggable.js";
 import {createIcon} from "../igv-icons.js";
-import { DomUtils } from '../../../node_modules/igv-ui/dist/igv-ui.js';
 
 const GenericContainer = function ({$parent, width, height, closeHandler}) {
